+++ c/include/sorting.h
@@ -8,6 +8,6 @@ static int partition(float *a, int left, int right, int pivot);
 
 static void quickselect(float *a, int left, int right, int k);
 
-float _percentile(const point* points, int number_of_points, float x_percent);
+float _percentile(const point* points, size_t number_of_points, float x_percent);
 
 #endif
